requirements.txt
py-cord
py-cord[voice]
youtube_dl
python-dotenv
emoji
